requirements.txt
pytest>=7.0
pydantic>=2.0
pydantic-settings>=2.0
fastapi>=0.95
uvicorn>=0.22
python-multipart>=0.0.5
deepgram-sdk>=3.0
openai>=1.0
langchain>=0.1
langchain-openai>=0.0

